Add templating for edit form
